see desc :raised_hands: :pizza:
made a check so having a sword in the weapon slot and swapping to a consumable in the inventory is prevented -> wrong slot items cannot be swapped if an item is already in the specific equipment slot. same id items can be stacked on to each other from equipment to inventory and vice versa. same item types can swapped into the equipment slots e.g sword (in inventory slot) can directly swap slots with a scimitar already in the sword slot.